refactor: moved mealdb API call to internal
