--- internal/mealdb_structs.go
@@ -1,10 +1,0 @@
-package internal
-
-type Meals struct {
-	Meal []Meal `json:"meals"`
-}
-
-type Meal struct {
-	Name string `json:"strMeal"`
-}
-
